ai/Agrovision-Chatbot-main/requirements.txt
# Natural Language Processing
google.generativeai
langchain-google-genai
langchain_community
python-dotenv
langchain
langchain-unstructured
chromadb
openai

# for visulization
requests
pillow
ipython

# for voice recognition
 SpeechRecognition
 faster_whisper
 pyttsx3
 pyaudio
pygame
langid
gtts

# to chat with our data
docx2txt
unstructured
unstructured[pdf]

#for api
flask